chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.129.0",
    "PyJWT>=2.8.0",
    "sqlalchemy>=2.0.46",
    "websockets>=16.0",
]
